Enhace lexer and parser
